Title :- Using Dental Metrics to Predict Gender.
Project Objective :-
To analyze the data based on a combination of dental features that describes the Gender of the person.

Methodology :-
1.) Data Collection
2.) Data Preprocessing
3.) Feature Engineering:
4.) Model Training and building

Conclusion :-
By leveraging dental metrics and advanced machine learning techniques, this project aims to provide a reliable method for predicting gender. The use of Logistic Regression, Decision Tree Classifier, Random forest Classifier and XGBoost ensures that the model is accurate and efficient, making it a valuable tool in forensic and anthropological studies.